# docs/swagger.yaml
definitions:
...
  useraccount.DriverInfo:
    properties:
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  useraccount.DriverLocation:
    properties:
      latitude:
        type: string
      longitude:
        type: string
    type: object
  useraccount.DriverStatus:
    properties:
      status:
        type: boolean
    type: object
  useraccount.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  useraccount.LoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  useraccount.NewAPIKeyResponse:
    properties:
      key:
        type: string
    type: object
  useraccount.UpdatePasswordRequest:
    properties:
      email:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    type: object
  useraccount.UserInfo:
    properties:
      email:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
    type: object
  useraccount.UserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - email
    - name
    - password
    - phone_number
    type: object
  useraccount.UserResponse:
    properties:
      createTime:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /apikey:
    get:
      consumes:
      - application/json
      description: Issues a user with a new api key, reissuing a new one invalidates
        old api keys
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      - description: email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Issues an API key
          schema:
            $ref: '#/definitions/useraccount.NewAPIKeyResponse'
        "400":
          description: Issues an API key
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Issues an API key
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: Issues an API key
...
  /driver/save-location:
    post:
      consumes:
      - application/json
      description: saves the location of a driver
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      - description: latitude and longitude
        in: body
        name: driverLocation
        required: true
        schema:
          $ref: '#/definitions/useraccount.DriverLocation'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Updates a job to assigned
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Updates a job to assigned
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: saves the location of a driver
  /driver/status:
    post:
      consumes:
      - application/json
      description: saves the status of a driver
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      - description: the status
        in: body
        name: driverLocation
        required: true
        schema:
          $ref: '#/definitions/useraccount.DriverStatus'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Updates a job to assigned
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Updates a job to assigned
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: saves the status of a driver
...
  /job/driver-info:
    get:
      consumes:
      - application/json
      description: Retrieves the information of a driver from an order
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      - description: job-id
        in: query
        name: job-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Gets the information of a driver from an order
          schema:
            $ref: '#/definitions/useraccount.DriverInfo'
        "400":
          description: Gets the information of a driver from an order
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Gets the information of a driver from an order
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: Retrieves the information of a driver from an order
...
  /login:
    post:
      consumes:
      - application/json
      description: Logins a user
      parameters:
      - description: Logins a user
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/useraccount.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: ""
          headers:
            Token:
              description: authorisation token
              type: string
        "400":
          description: Logins a user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Logins a user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: Logins a user
...
  /register/admin:
    post:
      consumes:
      - application/json
      description: Registers a new admin user
      parameters:
      - description: Registers a new user
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/useraccount.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registers a new user
          headers:
            Token:
              description: authorisation token
              type: string
          schema:
            $ref: '#/definitions/useraccount.UserResponse'
        "400":
          description: Registers a new user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Registers a new user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: Registers a new admin user
  /register/courier:
    post:
      consumes:
      - application/json
      description: Registers a new courier user
      parameters:
      - description: Registers a new user
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/useraccount.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registers a new user
          headers:
            Token:
              description: authorisation token
              type: string
          schema:
            $ref: '#/definitions/useraccount.UserResponse'
        "400":
          description: Registers a new user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Registers a new user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: Registers a new courier user
  /register/customer:
    post:
      consumes:
      - application/json
      description: Registers a new customer user
      parameters:
      - description: Registers a new user
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/useraccount.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registers a new user
          headers:
            Token:
              description: authorisation token
              type: string
          schema:
            $ref: '#/definitions/useraccount.UserResponse'
        "400":
          description: Registers a new user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Registers a new user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: Registers a new customer user
  /update-login:
    put:
      consumes:
      - application/json
      description: Updates a user password
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Logins a user
        in: body
        name: updatePassword
        required: true
        schema:
          $ref: '#/definitions/useraccount.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Logins a user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Logins a user
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: Updates a user password
  /user/info:
    get:
      consumes:
      - application/json
      description: Retrieves user information from the JWT token
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieves user information
          schema:
            $ref: '#/definitions/useraccount.UserInfo'
        "400":
          description: Retrieves user information
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
        "500":
          description: Retrieves user information
          schema:
            $ref: '#/definitions/useraccount.ErrorResponse'
      summary: Retrieves user information
...
swagger: "2.0"